Changed the BMP file loader to accept a std::string rather than a char array pointer. Reverted all the changes to memory handling in the Texture class until we can figure out why it's messing stuff up.

+++ src/drawing/texture/Texture.h
@@ -24,7 +24,7 @@ public:
 		};
 
 protected:
-	unsigned int width, height;
+	uint32_t width, height;
 	PixelFormat format;
 
 	void* pixelData;
@@ -32,22 +32,20 @@ protected:
 	Texture();	// so our subclasses can use this class
 
 public:
-	Texture(unsigned int newWidth, unsigned int newHeight, PixelFormat newFormat);
+	Texture(uint32_t newWidth, uint32_t newHeight, PixelFormat newFormat);
 	~Texture();
 
-	unsigned int getWidth() { return width; }
-	unsigned int getHeight() { return height; }
+	uint32_t getWidth() { return width; }
+	uint32_t getHeight() { return height; }
 	PixelFormat getFormat() { return format; }
 
 	void* getDataPointer() { return pixelData; }
 
-	unsigned short int getRedValueAt(unsigned int column, unsigned int row);
-	unsigned short int getGreenValueAt(unsigned int column, unsigned int row);
-	unsigned short int getBlueValueAt(unsigned int column, unsigned int row);
-	unsigned short int getAlphaValueAt(unsigned int column, unsigned int row);
-	void setColorAt(unsigned int column, unsigned int row, unsigned short int red, unsigned short int green, unsigned short int blue, unsigned short int alpha);
+	uint8_t getRedValueAt(uint32_t column, uint32_t row);
+	uint8_t getGreenValueAt(uint32_t column, uint32_t row);
+	uint8_t getBlueValueAt(uint32_t column, uint32_t row);
+	uint8_t getAlphaValueAt(uint32_t column, uint32_t row);
+	void setColorAt(uint32_t column, uint32_t row, uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha);
 };
 
 #endif
-
-
